refactor: moved all route routing logic to common routing
