+++ go/vt/vtgate/engine/delete.go
@@ -106,7 +106,7 @@ func (del *Delete) GetFields(VCursor, map[string]*querypb.BindVariable) (*sqltyp
 }
 
 func (del *Delete) execDeleteUnsharded(vcursor VCursor, bindVars map[string]*querypb.BindVariable) (*sqltypes.Result, error) {
-	rss, _, err := del.findRoutingInfo(vcursor, bindVars)
+	rss, _, err := del.findRoute(vcursor, bindVars)
 	if err != nil {
 		return nil, err
 	}
@@ -193,7 +193,7 @@ func (del *Delete) execDeleteIn(vcursor VCursor, bindVars map[string]*querypb.Bi
 }
 
 func (del *Delete) execDeleteByDestination(vcursor VCursor, bindVars map[string]*querypb.BindVariable, dest key.Destination) (*sqltypes.Result, error) {
-	rss, _, err := del.findRoutingInfo(vcursor, bindVars)
+	rss, _, err := del.findRoute(vcursor, bindVars)
 	if err != nil {
 		return nil, err
 	}
