+++ go/vt/vtgate/engine/routing.go
@@ -18,6 +18,7 @@ package engine
 
 import (
 	"encoding/json"
+	"strconv"
 
 	"vitess.io/vitess/go/sqltypes"
 	"vitess.io/vitess/go/vt/key"
@@ -118,8 +119,10 @@ type RoutingParameters struct {
 	Values []evalengine.Expr
 }
 
-func (rp *RoutingParameters) findRoutingInfo(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+func (rp *RoutingParameters) findRoute(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
 	switch rp.Opcode {
+	case None:
+		return nil, nil, nil
 	case DBA:
 		return rp.systemQuery(vcursor, bindVars)
 	case Unsharded, Next:
@@ -133,12 +136,24 @@ func (rp *RoutingParameters) findRoutingInfo(vcursor VCursor, bindVars map[strin
 	case Equal, EqualUnique:
 		switch rp.Vindex.(type) {
 		case vindexes.MultiColumn:
-			return rp.paramsSelectEqualMultiCol(vcursor, bindVars)
+			return rp.equalMultiCol(vcursor, bindVars)
 		default:
-			return rp.paramsSelectEqual(vcursor, bindVars)
+			return rp.equal(vcursor, bindVars)
 		}
-	case None:
-		return nil, nil, nil
+	case IN:
+		switch rp.Vindex.(type) {
+		case vindexes.MultiColumn:
+			return rp.inMultiCol(vcursor, bindVars)
+		default:
+			return rp.in(vcursor, bindVars)
+		}
+	case MultiEqual:
+		switch rp.Vindex.(type) {
+		case vindexes.MultiColumn:
+			return rp.paramsSelectMultiEqualMultiCol(vcursor, bindVars)
+		default:
+			return rp.paramsSelectMultiEqual(vcursor, bindVars)
+		}
 	default:
 		// Unreachable.
 		return nil, nil, vterrors.Errorf(vtrpcpb.Code_INTERNAL, "unsupported opcode: %v", rp.Opcode)
@@ -268,11 +283,6 @@ func (rp *RoutingParameters) routedTable(vcursor VCursor, bindVars map[string]*q
 	return nil, nil
 }
 
-func setReplaceSchemaName(bindVars map[string]*querypb.BindVariable) {
-	delete(bindVars, sqltypes.BvSchemaName)
-	bindVars[sqltypes.BvReplaceSchemaName] = sqltypes.Int64BindVariable(1)
-}
-
 func (rp *RoutingParameters) anyShard(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
 	rss, _, err := vcursor.ResolveDestinations(rp.Keyspace.Name, nil, []key.Destination{key.DestinationAnyShard{}})
 	if err != nil {
@@ -312,7 +322,7 @@ func (rp *RoutingParameters) byDestination(vcursor VCursor, bindVars map[string]
 	return rss, multiBindVars, err
 }
 
-func (rp *RoutingParameters) paramsSelectEqual(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+func (rp *RoutingParameters) equal(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
 	env := evalengine.EnvWithBindVars(bindVars)
 	value, err := env.Evaluate(rp.Values[0])
 	if err != nil {
@@ -329,7 +339,7 @@ func (rp *RoutingParameters) paramsSelectEqual(vcursor VCursor, bindVars map[str
 	return rss, multiBindVars, nil
 }
 
-func (rp *RoutingParameters) paramsSelectEqualMultiCol(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+func (rp *RoutingParameters) equalMultiCol(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
 	env := evalengine.EnvWithBindVars(bindVars)
 	var rowValue []sqltypes.Value
 	for _, rvalue := range rp.Values {
@@ -349,4 +359,192 @@ func (rp *RoutingParameters) paramsSelectEqualMultiCol(vcursor VCursor, bindVars
 		multiBindVars[i] = bindVars
 	}
 	return rss, multiBindVars, nil
+}
+
+func (rp *RoutingParameters) in(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+	env := evalengine.EnvWithBindVars(bindVars)
+	value, err := env.Evaluate(rp.Values[0])
+	if err != nil {
+		return nil, nil, err
+	}
+	rss, values, err := resolveShards(vcursor, rp.Vindex.(vindexes.SingleColumn), rp.Keyspace, value.TupleValues())
+	if err != nil {
+		return nil, nil, err
+	}
+	return rss, shardVars(bindVars, values), nil
+}
+
+func (rp *RoutingParameters) inMultiCol(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+	rowColValues, isSingleVal, err := generateRowColValues(bindVars, rp.Values)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	rss, mapVals, err := resolveShardsMultiCol(vcursor, rp.Vindex.(vindexes.MultiColumn), rp.Keyspace, rowColValues, true /* shardIdsNeeded */)
+	if err != nil {
+		return nil, nil, err
+	}
+	return rss, shardVarsMultiCol(bindVars, mapVals, isSingleVal), nil
+}
+
+func (rp *RoutingParameters) paramsSelectMultiEqual(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+	env := evalengine.EnvWithBindVars(bindVars)
+	value, err := env.Evaluate(rp.Values[0])
+	if err != nil {
+		return nil, nil, err
+	}
+	rss, _, err := resolveShards(vcursor, rp.Vindex.(vindexes.SingleColumn), rp.Keyspace, value.TupleValues())
+	if err != nil {
+		return nil, nil, err
+	}
+	multiBindVars := make([]map[string]*querypb.BindVariable, len(rss))
+	for i := range multiBindVars {
+		multiBindVars[i] = bindVars
+	}
+	return rss, multiBindVars, nil
+}
+
+func (rp *RoutingParameters) paramsSelectMultiEqualMultiCol(vcursor VCursor, bindVars map[string]*querypb.BindVariable) ([]*srvtopo.ResolvedShard, []map[string]*querypb.BindVariable, error) {
+	var multiColValues [][]sqltypes.Value
+	env := evalengine.EnvWithBindVars(bindVars)
+	for _, rvalue := range rp.Values {
+		v, err := env.Evaluate(rvalue)
+		if err != nil {
+			return nil, nil, err
+		}
+		multiColValues = append(multiColValues, v.TupleValues())
+	}
+
+	// transpose from multi col value to vindex keys with one value from each multi column values.
+	// [1,3]
+	// [2,4]
+	// [5,6]
+	// change
+	// [1,2,5]
+	// [3,4,6]
+
+	rowColValues := make([][]sqltypes.Value, len(multiColValues[0]))
+	for _, colValues := range multiColValues {
+		for row, colVal := range colValues {
+			rowColValues[row] = append(rowColValues[row], colVal)
+		}
+	}
+
+	rss, _, err := resolveShardsMultiCol(vcursor, rp.Vindex.(vindexes.MultiColumn), rp.Keyspace, rowColValues, false /* shardIdsNotNeeded */)
+	if err != nil {
+		return nil, nil, err
+	}
+	multiBindVars := make([]map[string]*querypb.BindVariable, len(rss))
+	for i := range multiBindVars {
+		multiBindVars[i] = bindVars
+	}
+	return rss, multiBindVars, nil
+}
+
+func setReplaceSchemaName(bindVars map[string]*querypb.BindVariable) {
+	delete(bindVars, sqltypes.BvSchemaName)
+	bindVars[sqltypes.BvReplaceSchemaName] = sqltypes.Int64BindVariable(1)
+}
+
+func resolveShards(vcursor VCursor, vindex vindexes.SingleColumn, keyspace *vindexes.Keyspace, vindexKeys []sqltypes.Value) ([]*srvtopo.ResolvedShard, [][]*querypb.Value, error) {
+	// Convert vindexKeys to []*querypb.Value
+	ids := make([]*querypb.Value, len(vindexKeys))
+	for i, vik := range vindexKeys {
+		ids[i] = sqltypes.ValueToProto(vik)
+	}
+
+	// Map using the Vindex
+	destinations, err := vindex.Map(vcursor, vindexKeys)
+	if err != nil {
+		return nil, nil, err
+
+	}
+
+	// And use the Resolver to map to ResolvedShards.
+	return vcursor.ResolveDestinations(keyspace.Name, ids, destinations)
+}
+
+func resolveShardsMultiCol(vcursor VCursor, vindex vindexes.MultiColumn, keyspace *vindexes.Keyspace, rowColValues [][]sqltypes.Value, shardIdsNeeded bool) ([]*srvtopo.ResolvedShard, [][][]*querypb.Value, error) {
+	destinations, err := vindex.Map(vcursor, rowColValues)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	// And use the Resolver to map to ResolvedShards.
+	rss, shardsValues, err := vcursor.ResolveDestinationsMultiCol(keyspace.Name, rowColValues, destinations)
+	if err != nil {
+		return nil, nil, err
+	}
+
+	if shardIdsNeeded {
+		return rss, buildMultiColumnVindexValues(shardsValues), nil
+	}
+	return rss, nil, nil
+}
+
+// buildMultiColumnVindexValues takes in the values resolved for each shard and transposes them
+// and eliminates duplicates, returning the values to be used for each column for a multi column
+// vindex in each shard.
+func buildMultiColumnVindexValues(shardsValues [][][]sqltypes.Value) [][][]*querypb.Value {
+	var shardsIds [][][]*querypb.Value
+	for _, shardValues := range shardsValues {
+		// shardValues -> [[0,1], [0,2], [0,3]]
+		// shardIds -> [[0,0,0], [1,2,3]]
+		// cols = 2
+		cols := len(shardValues[0])
+		shardIds := make([][]*querypb.Value, cols)
+		colValSeen := make([]map[string]interface{}, cols)
+		for _, values := range shardValues {
+			for colIdx, value := range values {
+				if colValSeen[colIdx] == nil {
+					colValSeen[colIdx] = map[string]interface{}{}
+				}
+				if _, found := colValSeen[colIdx][value.String()]; found {
+					continue
+				}
+				shardIds[colIdx] = append(shardIds[colIdx], sqltypes.ValueToProto(value))
+				colValSeen[colIdx][value.String()] = nil
+			}
+		}
+		shardsIds = append(shardsIds, shardIds)
+	}
+	return shardsIds
+}
+
+func shardVars(bv map[string]*querypb.BindVariable, mapVals [][]*querypb.Value) []map[string]*querypb.BindVariable {
+	shardVars := make([]map[string]*querypb.BindVariable, len(mapVals))
+	for i, vals := range mapVals {
+		newbv := make(map[string]*querypb.BindVariable, len(bv)+1)
+		for k, v := range bv {
+			newbv[k] = v
+		}
+		newbv[ListVarName] = &querypb.BindVariable{
+			Type:   querypb.Type_TUPLE,
+			Values: vals,
+		}
+		shardVars[i] = newbv
+	}
+	return shardVars
+}
+
+func shardVarsMultiCol(bv map[string]*querypb.BindVariable, mapVals [][][]*querypb.Value, isSingleVal map[int]interface{}) []map[string]*querypb.BindVariable {
+	shardVars := make([]map[string]*querypb.BindVariable, len(mapVals))
+	for i, shardVals := range mapVals {
+		newbv := make(map[string]*querypb.BindVariable, len(bv)+len(shardVals)-len(isSingleVal))
+		for k, v := range bv {
+			newbv[k] = v
+		}
+		for j, vals := range shardVals {
+			if _, found := isSingleVal[j]; found {
+				// this vindex column is non-tuple column hence listVal bind variable is not required to be set.
+				continue
+			}
+			newbv[ListVarName+strconv.Itoa(j)] = &querypb.BindVariable{
+				Type:   querypb.Type_TUPLE,
+				Values: vals,
+			}
+		}
+		shardVars[i] = newbv
+	}
+	return shardVars
 }
